+++ lib/src/commands/types/mp_command_description_type.dart
@@ -68,6 +68,7 @@ enum MPCommandDescriptionType {
   simplifyStraight,
   simplifyToBezier,
   simplifyToStraight,
+  splitLineAtSelectedPoints,
   toggleReverseOption,
   toggleSmoothOption;