+++ lib/src/controllers/th2_file_edit_controller.dart
@@ -18,6 +18,7 @@ import 'package:mapiah/src/controllers/mp_visual_controller.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_area_line_creation_controller.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_copy_paste_controller.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_element_edit_controller.dart';
+import 'package:mapiah/src/controllers/th2_file_edit_split_merge_controller.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_option_edit_controller.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_search_controller.dart';
 import 'package:mapiah/src/controllers/th2_file_hide_element_controller.dart';
@@ -57,6 +58,7 @@ abstract class TH2FileEditControllerBase with Store {
   late final MPVisualController visualController;
   late final TH2FileEditCopyPasteController copyPasteController;
   late final TH2FileEditElementEditController elementEditController;
+  late final TH2FileEditSplitMergeController splitMergeController;
   late final TH2FileEditOptionEditController optionEditController;
   late final TH2FileEditOverlayWindowController overlayWindowController;
   late final TH2FileEditSearchController searchController;
@@ -458,6 +460,9 @@ abstract class TH2FileEditControllerBase with Store {
       this as TH2FileEditController,
     );
     elementEditController = TH2FileEditElementEditController(
+      this as TH2FileEditController,
+    );
+    splitMergeController = TH2FileEditSplitMergeController(
       this as TH2FileEditController,
     );
     optionEditController = TH2FileEditOptionEditController(